refactor(runtime): retrofit CLI gazetteer + overlay plugin onto WorldZonesRuntime.Build
Delete the two surviving copies of the region bootstrap, routing both consumers
through the shared façade from 82ed43b. Behaviour-preserving — proven, not asserted.

CLI gazetteer (Gazetteer.cs):
- Export() now calls WorldZonesRuntime.Build instead of hand-rolling classify → label
  → GenerateLand → aggregate. The private Agg struct + per-zone aggregation loop are gone
  (lifted into GazetteerBuilder). Writers consume RegionInfo.
- Names pinned to LegacyRegionNamer so output is byte-identical (the richer multi-schema
  namer is a separate, visible opt-in — not smuggled into a refactor).
- VERIFIED byte-identical vs a pre-retrofit baseline: TSV + grid.bin exact, JSON exact
  except the generatedUtc timestamp. Caught + fixed two drift bugs via the diff: biome
  fractions must divide by sampled-zone count (added RegionInfo.SampledLandZones), and TSV
  row order is RegionKey not area.

Overlay plugin (RegionOverlayPlugin.cs):
- TryInitializeRegionData() now calls Build with ComputeRegionInfo=false — a new option that
  skips the rich aggregation + naming for point-query-only consumers, so this path stays
  height-only (no biome bridge) exactly like before. New ValheimWorldSampler implements
  IWorldSampler against the live WorldGenerator.
- worldId KEPT as the numeric seed string: discovery-persistence file paths key on it, so
  changing it would orphan saved discovery state. (The in-game-vs-dataset name skew is a
  separate deliberate decision, not a thing this refactor flips.)
- Dead ValheimWorldDataProvider.cs removed.
- COMPILE-VERIFIED on Linux via VALHEIM_MODDED_PATH (merged client+BepInEx ref root):
  Build succeeded, 0 errors. Runtime (in-world) still needs a real client — unchanged.

Model refinements (RegionInfo): centroid/mean now double (averaged fields; matches the
gazetteer's double-then-round); BiomeZoneCounts added (lossless); RegionAt gained a double
overload (consumers querying at a region centroid no longer need a cast).

Regression tripwire: 73 Regions + 46 WorldGen tests green. Consumer demo green, names
unchanged (162/162 unique).

+++ src/WorldZones.Runtime/MultiSchemaRegionNamer.cs
@@ -164,7 +164,7 @@ namespace WorldZones.Runtime
         }
 
         // ── cardinal direction + radial distance from world centre ──
-        private static (string card, double dist) Cardinal(float cx, float cz)
+        private static (string card, double dist) Cardinal(double cx, double cz)
         {
             double ang = Math.Atan2(cz, cx) * 180.0 / Math.PI;
             // (name, lo, hi) — first matching half-open [lo,hi) wins; Vest wraps the ±180 seam.
